Rename Expansion of Jimmy Ng Tidyverse by Samantha Deokinanan.Rmd to Extension By Yohannes Deboch to Jimmy and Samantha
Hi Jimmy and Samantha . Thank you for sharing . The Tidy receipts by Jimmmy and the extension by Samantha are impressive . After carefully analyzing the work ; I felt like adding some inputs on the conclusion section . 

Conclusion 
ly checking availbale options I have decided to add an input on the conclusion section . The provided input is based on my personal insights and codes submitted above Jimmy and Samanatha. Dplyr would benefit from having a function summarizing the data frame variables. It is surprising that the R base package has nothing better than the summary function to provide an overview of a data frame. In dplyr one can look at the data with for example glimpse or head, but a concise display of key summary statistics would make data management easier. Summary statistics can provide more information than the raw data.The dplyr package is an excellent addition to the  workflow. It speeds up, cleans up and simplifies data analysis and manipulation.
In the end, `dplyr` provides multiple functions that can be used when you want to create an output value that is conditioned on an evaluation. The right function can make the transformation less cumbersome.